Move user search/validation to own component
